+++ R/type_convert.R
@@ -7,6 +7,7 @@
 #' @param df A data frame.
 #' @inheritParams col_types_standardise
 #' @inheritParams tokenizer_delim
+#' @inheritParams read_csv
 #' @export
 #' @examples
 #' df <- data.frame(
@@ -32,7 +33,7 @@ type_convert <- function(df, col_types = NULL, na = c("", "NA"), trim_ws = TRUE,
 
   df[is_character] <- lapply(seq_along(char_cols), function(i) {
     type_convert_col(char_cols[[i]], col_types[[i]], which(is_character)[i],
-      na = na, trim_ws = trim_ws)
+      locale_ = locale, na = na, trim_ws = trim_ws)
   })
 
   df